Make filename soudly
