+++ demo/strutures/point.go
