+++ demo/strutures/point_test.go
