+++ demo/virtual_machine/main.go
